WIP: a little clean up
